Main and Config classes improved

+++ src/main/java/com/example/berkdeneme/ConfigController.java
@@ -1,5 +1,6 @@
 package com.example.berkdeneme;
 
+import com.example.berkdeneme.Compilers.CCompiler;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
@@ -12,11 +13,13 @@ import javafx.scene.control.TextField;
 import javafx.stage.FileChooser;
 
 import java.io.File;
+import java.io.IOException;
 import java.io.ObjectInputFilter;
 import java.net.URL;
 import java.nio.file.Files;
 import java.util.Objects;
 import java.util.ResourceBundle;
+import static com.example.berkdeneme.Main.stage;
 
 public class ConfigController implements Initializable {
     @FXML
@@ -54,25 +57,33 @@ public class ConfigController implements Initializable {
         list.addAll("C","Java","Python");
         configComboBox.setItems(list);
         configComboBox.getSelectionModel().selectFirst();
-        /*if (ObjectInputFilter.Config.getInstance()){
-            compilerPath.setText();
-            assignmentPath.setText();
-            args.setText();
-            runField.setText();
-            expected.setText();
-            Language
-      //  }*/
+        if (Config.getInstance().compilerPath != null){
+            compilerPath.setText(Config.getInstance().compilerPath);
+            assignmentPath.setText(Config.getInstance().assignmentPath);
+            args.setText(Config.getInstance().args);
+            runField.setText(Config.getInstance().run);
+            expected.setText(Config.getInstance().expected);
+            Language selectedLang = Config.getInstance().selectedLanguage;
+            if (selectedLang == null){
+            }else if (selectedLang.toString().equals("C")){
+                configComboBox.getSelectionModel().select(1);
+            }else if (selectedLang.toString().equals("Java")){
+                configComboBox.getSelectionModel().select(2);
+            }else if (selectedLang.toString().equals("Python")){
+                configComboBox.getSelectionModel().select(3);
+            }
+        }
         runButton.setOnAction(actionEvent -> {
             try {
                 if (compilerPath.getText() == null || args.getText() == null || assignmentPath.getText() == null || expected.getText() == null || runField.getText() == null ||compilerPath.getText().isBlank() ||assignmentPath.getText().isBlank() || args.getText().isBlank() || runField.getText().isBlank() || expected.getText().isBlank()){
                     throw new Exception("There is a field not filled.");
                 }
-                /*
-                Config
-                Config
-                Config
-
-                 */
+                Config.getInstance().compilerPath = compilerPath.getText();
+                Config.getInstance().assignmentPath = assignmentPath.getText();
+                Config.getInstance().args = args.getText();
+                Config.getInstance().run = runField.getText();
+                Config.getInstance().expected = expected.getText();
+                Config.getInstance().runArgs = runFieldArgs.getText();
             }catch (Exception e){
                 e.printStackTrace();
 
@@ -82,67 +93,62 @@ public class ConfigController implements Initializable {
         StringBuilder selectedDirectory = new StringBuilder();
         assignmentImportButton.setOnAction(actionEvent -> {
             try {
-               /* if (assignmentPath != null){
-
-                }*/
-
+                configImport(selectedDirectory,list);
+                if (assignmentPath != null){
+                    Config.getInstance().assignmentPath = assignmentPath.getText();
+                }
             }catch (Exception e){
                 e.printStackTrace();
             }
         });
       runImportButton.setOnAction(actionEvent -> {
           FileChooser fc = new FileChooser();
-         // File selectedF = fc.showOpenDialog()
-          //if ()
-          try {
-            // String content = Files.readString();
-            //runFieldArgs.setText();
-          }catch (Exception e){
-              e.printStackTrace();
+          File selectedF = fc.showOpenDialog(stage);
+          if (selectedF != null) {
+              try {
+                  String content = Files.readString(selectedF.toPath());
+                  runFieldArgs.setText(content);
+              } catch (IOException e) {
+                  e.printStackTrace();
+              }
           }
       });
       expectedImportButton.setOnAction(actionEvent -> {
           FileChooser fc = new FileChooser();
-          // File selectedF = fc.showOpenDialog()
-          //if ()
-          try {
-              // String content = Files.readString();
-              //runFieldArgs.setText();
-          }catch (Exception e){
-              e.printStackTrace();
+          File selectedF = fc.showOpenDialog(stage);
+          if (selectedF != null) {
+              try {
+                  String content = Files.readString(selectedF.toPath());
+                  expected.setText(content);
+              } catch (IOException e) {
+                  e.printStackTrace();
+              }
           }
       });
 
-      //saveButton.setOnAction(actionEvent -> );
+      saveButton.setOnAction(actionEvent -> configExport(selectedDirectory));
         configComboBox.valueProperty().addListener(new ChangeListener<String>() {
             @Override
             public void changed(ObservableValue<? extends String> observableValue, String s, String t1) {
                 String comboBoxVal = configComboBox.getValue();
-             /* String directoryStr = Config;
+              String directoryStr = Config.getInstance().assignmentPath;
                 if (comboBoxVal == null) {
                 } else if (comboBoxVal.equals("C")) {
-
+                    Config.getInstance().selectedLanguage = Language.C;
                 } else if (comboBoxVal.equals("Java")) {
-
+                    Config.getInstance().selectedLanguage = Language.JAVA;
                 } else if (comboBoxVal.equals("Python")) {
-
-                }*/
+                    Config.getInstance().selectedLanguage = Language.PYTHON;
+                }
             }
         });
 
         deleteButton.setOnAction(actionEvent -> {
             String filePath = selectedDirectory + File.separator + "config";
             File f = new File(filePath);
-           /* if (f.exists()){
+            if (f.exists()) {
                 boolean deleted = f.delete();
-                if (deleted){
-
-                }else {
-
-                }
-            }else {
-
-            }*/
+            }
         });
     }
 
@@ -157,7 +163,7 @@ public class ConfigController implements Initializable {
         runImportButton.getParent().setVisible(runFieldBool);
         runImportButton.getParent().setManaged(runFieldBool);
         runField.setText(runFieldPm);
-        //expected
+        expected.setText(Config.getInstance().expected);
     }
     private void configImport(){
